+++ bash/decoders/vdlm2dec.sh
@@ -7,18 +7,18 @@ source $RECEIVER_BASH_DIRECTORY/functions.sh
 
 clear
 log_project_title
-log_title_heading "Setting up the ACARSDEC decoder"
+log_title_heading "Setting up the VDLM2DEC decoder"
 log_title_message "------------------------------------------------------------------------------"
 if ! whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
-              --title "ACARSDEC decoder Setup" \
-              --yesno "ACARSDEC is a multi-channels acars decoder with built-in rtl_sdr, airspy front end or sdrplay device. Since 3.0, It comes with a database backend : acarsserv to store received acars messages.\n\nWould you like to begin the setup process now?" \
+              --title "VDLM2DEC decoder Setup" \
+              --yesno "VDLM2DEC is a vdl mode 2 decoder with built-in rtl_sdr or airspy front end.\n\nWould you like to begin the setup process now?" \
               11 78; then
     echo ""
     log_alert_heading "INSTALLATION HALTED"
     log_alert_message "Setup has been halted at the request of the user"
     echo ""
     log_title_message "------------------------------------------------------------------------------"
-    log_title_heading "ACARSDEC decoder setup halted"
+    log_title_heading "VDLM2DEC decoder setup halted"
     echo ""
     exit 1
 fi
@@ -26,16 +26,15 @@ fi
 
 ## GATHER REQUIRED INFORMATION FROM THE USER
 
-log_heading "Determine the device type to build ACARSDEC for"
+log_heading "Determine the device type to build VDLM2DEC for"
 
-log_message "Asking which type of device will be used by ACARSDEC"
+log_message "Asking which type of device will be used by VDLM2DEC"
 device=$(whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
                   --title "Device Type" \
-                  --menu "Please choose the RTL-SDR device type which is to be used by ACARSDEC." \
+                  --menu "Please choose the RTL-SDR device type which is to be used by VDLM2DEC." \
                   11 78 3 \
                   "RTL-SDR" "" \
                   "AirSpy" "" \
-                  "SDRPlay" "" \
                   3>&1 1>&2 2>&3)
 exit_status=$?
 if [[ $exit_status != 0 ]]; then
@@ -44,64 +43,56 @@ if [[ $exit_status != 0 ]]; then
     log_alert_message "Setup has been halted at the request of the user"
     echo ""
     log_title_message "------------------------------------------------------------------------------"
-    log_title_heading "ACARSDEC decoder setup halted"
+    log_title_heading "VDLM2DEC decoder setup halted"
     echo ""
     exit 1
 fi
 
-ask_for_device_assignments "acarsdec"
+ask_for_device_assignments "vdlm2dec"
 if [[ $? -ne 0 ]] ; then
     log_alert_heading "INSTALLATION HALTED"
     log_alert_message "Setup has been halted due to lack of required information"
     echo ""
     log_title_message "------------------------------------------------------------------------------"
-    log_title_heading "ACARSDEC decoder setup halted"
+    log_title_heading "VDLM2DEC decoder setup halted"
     exit 1
 fi
-if [[ -z $RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER ]]; then
-    RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER="0"
-fi
 
-current_acars_frequencies="130.025 130.425 130.450 131.125 131.550"
-if [[ "${acars_decoder_installed}" == "true" ]]; then
+current_vdlm2_frequencies="136.100 136.650 136.700 136.800 136.975"
+if [[ -f /etc/systemd/system/dumpvdl2.service ]]; then
     log_message "Determining which frequencies are currently assigned"
-    exec_start=`get_config "ExecStart" "/etc/systemd/system/acarsdec.service"`
-    current_acars_frequencies=`sed -e "s#.*-r ${RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER} \(\)#\1#" <<< "${exec_start}"`
+    exec_start=`get_config "ExecStart" "/etc/systemd/system/vdlm2dec.service"`
+    current_vdlm2_frequencies=`sed -e "s#.*-r ${RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER} \(\)#\1#" <<< "${exec_start}"`
 fi
-log_message "Asking the user for ACARS frequencies to monitor"
-acars_fequencies_title="Enter ACARS Frequencies"
-while [[ -z $acars_fequencies ]] ; do
-    acars_fequencies=$(whiptail --backtitle "ACARS Frequencies" \
-                                --title "${acars_fequencies_title}" \
-                                --inputbox "\nEnter the ACARS frequencies you would like to monitor." \
-                                8 78 \
-                                "${current_acars_frequencies}" 3>&1 1>&2 2>&3)
+log_message "Asking the user for VDL Mode 2 frequencies to monitor"
+vdlm2_fequencies_title="Enter VDL Mode 2 Frequencies"
+while [[ -z $vdlm2_fequencies ]] ; do
+    vdlm2_fequencies=$(whiptail --backtitle "VDL Mode 2 Frequencies" \
+                              --title "${vdlm2_fequencies_title}" \
+                              --inputbox "\nEnter the VDL Mode 2 frequencies you would like to monitor." \
+                              8 78 \
+                              "${current_vdlm2_frequencies}" 3>&1 1>&2 2>&3)
     exit_status=$?
     if [[ $exit_status != 0 ]]; then
         log_alert_heading "INSTALLATION HALTED"
         log_alert_message "Setup has been halted due to lack of required information"
         echo ""
         log_title_message "------------------------------------------------------------------------------"
-        log_title_heading "ACARSDEC decoder setup halted"
+        log_title_heading "VDLM2DEC decoder setup halted"
         exit 1
     fi
-    acars_fequencies_title="Enter ACARS Frequencies (REQUIRED)"
+    vdlm2_fequencies_title="Enter VDL Frequencies (REQUIRED)"
 done
 
 
 ## CHECK FOR PREREQUISITE PACKAGES
 
-log_heading "Installing packages needed to fulfill dependencies for FlightAware PiAware client"
-
 check_package cmake
 check_package libjansson-dev
-check_package libpaho-mqtt-dev
-check_package libsndfile1-dev
 check_package libsqlite3-dev
-check_package libusb-1.0-0-dev
 check_package libxml2-dev
-check_package pkg-config
 check_package zlib1g-dev
+
 
 case $RECEIVER_OS_DISTRIBUTION in
     ubuntu)
@@ -114,18 +105,6 @@ case $RECEIVER_OS_DISTRIBUTION in
 esac
 check_package sqlite3
 check_package php${distro_php_version}-sqlite3
-
-case "${device}" in
-    "RTL-SDR")
-        check_package librtlsdr-dev
-        ;;
-    "AirSpy")
-        check_package libairspy-dev
-        ;;
-    "SDRPlay")
-        check_package libmirisdr-dev
-        ;;
-esac
 
 
 ## BLACKLIST UNWANTED RTL-SDR MODULES
@@ -186,41 +165,41 @@ log_message "Running ldconfig"
 sudo ldconfig
 
 
-## CLONE OR PULL THE ACARSDEC GIT REPOSITORY
+## CLONE OR PULL THE VDLM2DEC GIT REPOSITORY
 
-log_heading "Preparing the ACARSDEC Git repository"
+log_heading "Preparing the VDLM2DEC Git repository"
 
-if [[ -d $RECEIVER_BUILD_DIRECTORY/acarsdec && -d $RECEIVER_BUILD_DIRECTORY/acarsdec/.git ]]; then
-    log_message "Entering the ACARSDEC git repository directory"
-    cd $RECEIVER_BUILD_DIRECTORY/acarsdec
-    log_message "Updating the local ACARSDEC git repository"
+if [[ -d $RECEIVER_BUILD_DIRECTORY/vdlm2dec && -d $RECEIVER_BUILD_DIRECTORY/vdlm2dec/.git ]]; then
+    log_message "Entering the VDLM2DEC git repository directory"
+    cd $RECEIVER_BUILD_DIRECTORY/vdlm2dec
+    log_message "Updating the local VDLM2DEC git repository"
     echo ""
     git pull
 else
     log_message "Entering the build directory"
     cd $RECEIVER_BUILD_DIRECTORY
-    log_message "Cloning the ACARSDEC git repository locally"
+    log_message "Cloning the VDLM2DEC git repository locally"
     echo ""
-    git clone https://github.com/TLeconte/acarsdec.git
+    git clone https://github.com/TLeconte/vdlm2dec.git
 fi
 
 
-## BUILD AND INSTALL THE ACARSDEC BINARY
+## BUILD AND INSTALL THE VDLM2DEC BINARY
 
-log_heading "Building the ACARSDEC binary"
+log_heading "Building the VDLM2DEC binary"
 
-if [[ ! -d $RECEIVER_BUILD_DIRECTORY/acarsdec/build ]]; then
-    log_message "Creating the ACARSDEC build directory"
+if [[ ! -d $RECEIVER_BUILD_DIRECTORY/vdlm2dec/build ]]; then
+    log_message "Creating the VDLM2DEC build directory"
     echo ""
-        mkdir -vp $RECEIVER_BUILD_DIRECTORY/acarsdec/build
+        mkdir -vp $RECEIVER_BUILD_DIRECTORY/vdlm2dec/build
     echo ""
 fi
-if [[ -n "$(ls -A $RECEIVER_BUILD_DIRECTORY/acarsdec/build 2>/dev/null)" ]]; then
-    log_message "Deleting all files currently residing in the ACARSDEC build directory"
-    rm -rf $RECEIVER_BUILD_DIRECTORY/acarsdec/build/*
+if [[ -n "$(ls -A $RECEIVER_BUILD_DIRECTORY/vdlm2dec/build 2>/dev/null)" ]]; then
+    log_message "Deleting all files currently residing in the VDLM2DEC build directory"
+    rm -rf $RECEIVER_BUILD_DIRECTORY/vdlm2dec/build/*
 fi
-log_message "Entering the ACARSDEC build directory"
-cd $RECEIVER_BUILD_DIRECTORY/acarsdec/build
+log_message "Entering the VDLM2DEC build directory"
+cd $RECEIVER_BUILD_DIRECTORY/vdlm2dec/build
 
 log_message "Executing cmake"
 echo ""
@@ -230,9 +209,6 @@ case "${device}" in
         ;;
     "AirSpy")
         cmake .. -Dairspy=ON
-        ;;
-    "SDRPlay")
-        cmake .. -Dsdrplay=ON
         ;;
 esac
 echo ""
@@ -277,16 +253,16 @@ make -f Makefile
 echo ""
 
 
-## RUN ACARSDECO AND ACARSSERV
+## RUN VDLM2DEC AND ACARSSERV
 
-log_message "Creating the ACARSDEC systemd service script"
-sudo tee /etc/systemd/system/acarsdec.service > /dev/null <<EOF
+log_message "Creating the VDLM2DEC systemd service script"
+sudo tee /etc/systemd/system/vdlm2dec.service > /dev/null <<EOF
 [Unit]
 Description=ARCARSDEC multi-channel acars decoder.
 After=network.target
 
 [Service]
-ExecStart=/usr/local/bin/acarsdec -j 127.0.0.1:5555 -o2 -g280 -r 0 130.025 130.425 130.450 131.125 131.550
+ExecStart=/usr/local/bin/vdlm2dec -j 127.0.0.1:5555 -G -g280 -r 0 ${current_vdlm2_frequencies}
 WorkingDirectory=/usr/local/bin
 StandardOutput=null
 TimeoutSec=30
@@ -319,8 +295,8 @@ StartLimitBurst=10
 WantedBy=multi-user.target
 EOF
 
-log_message "Enabling then starting the ACARSDEC service"
-sudo systemctl enable --now acarsdec.service
+log_message "Enabling then starting the VDLM2DEC service"
+sudo systemctl enable --now vdlm2dec.service
 log_message "Enabling then starting the acarsserv service"
 sudo systemctl enable --now acarsserv.service
 
@@ -333,8 +309,8 @@ assign_devices_to_decoders
 ## POST INSTALLATION OPERATIONS
 
 whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
-         --title "ACARSDEC Decoder Setup Complete" \
-         --msgbox "The setup process currently sets basic parameters needed to feed acarsserv. You can fine tune your installation by modifying the startup command found in the file /etc/systemd/system/acarsdec.service. Usage information for ACARSDEC can be found in the projects README at https://github.com/TLeconte/acarsdec." \
+         --title "VDLM2DEC Decoder Setup Complete" \
+         --msgbox "The setup process currently sets basic parameters needed to feed acarsserv. You can fine tune your installation by modifying the startup command found in the file /etc/systemd/system/vdlm2dec.service. Usage information for VDLM2DEC can be found in the projects README at https://github.com/TLeconte/vdlm2dec." \
          12 78
 
 
@@ -345,7 +321,7 @@ cd $RECEIVER_ROOT_DIRECTORY
 
 echo ""
 log_title_message "------------------------------------------------------------------------------"
-log_title_heading "ACARSDEC decoder setup is complete"
+log_title_heading "VDLM2DEC decoder setup is complete"
 echo ""
 read -p "Press enter to continue..." discard