+++ bash/main.sh
@@ -8,11 +8,13 @@ source ${RECEIVER_BASH_DIRECTORY}/functions.sh
 
 ## ADS-B DECODERS
 
-# FlightAware dump1090
+adsb_decoder_installed="false"
 install_adsb_decoder="false"
-if [[ $(dpkg-query -W -f='${STATUS}' dump1090-fa 2>/dev/null | grep -c "ok installed") == 1 ]] ; then
+
+if [[ $(dpkg-query -W -f='${STATUS}' dump1090-fa 2>/dev/null | grep -c "ok installed") == 1 ]]; then
+    adsb_decoder_installed="true"
     chosen_adsb_decoder="dump1090-fa"
-    if [[ $(sudo dpkg -s dump1090-fa 2>/dev/null | grep -c "Version: ${dump1090_fa_current_version}") == 0 ]] ; then
+    if [[ $(sudo dpkg -s dump1090-fa 2>/dev/null | grep -c "Version: ${dump1090_fa_current_version}") == 0 ]]; then
         whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
                  --title "FlightAware Dump1090 Upgrade Available" \
                  --defaultno \
@@ -30,15 +32,34 @@ if [[ $(dpkg-query -W -f='${STATUS}' dump1090-fa 2>/dev/null | grep -c "ok insta
             install_adsb_decoder="true"
         fi
     fi
-else
+fi
+
+if [[ $(dpkg-query -W -f='${STATUS}' readsb 2>/dev/null | grep -c "ok installed") == 1 ]]; then
+    adsb_decoder_installed="true"
+    chosen_adsb_decoder="readsb"
     whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
-             --title "FlightAware dump978" \
+             --title "Reinstall Readsb Decoder" \
              --defaultno \
-             --yesno "FlightAware dump1090 is capable of demodulating ADS-B, Mode S, Mode 3A/3C signals received by an SDR device.\n\nGitHub Repository: https://github.com/flightaware/dump1090\n\nWould you like to install FlightAware dump1090?" \
-             10 65
+             --yesno "The option to rebuild and reinstall Readsb is available.\n\nWould you like to rebuild and reinstall Readsb?" \
+             9 65
     if [[ $? == 0 ]]; then
         install_adsb_decoder="true"
-        chosen_adsb_decoder="dump1090-fa"
+    fi
+fi
+
+if [[ "${adsb_decoder_installed}" == "false" ]]; then
+    install_adsb_decoder="true"
+    chosen_adsb_decoder=$(whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
+                                   --title "ADS-B Decoder Selection" \
+                                   --menu "The following ADS-B decoders are available for installation." \
+                                   16 100 9 \
+                                   "None" "Do not install an ADS-B decoder." \
+                                   "dump1090-fa" "FlightAware's version of the dump1090 decoder." \
+                                   "readsb" "Wiedehopf's detached fork of readsb." \
+                                   3>&2 2>&1 1>&3)
+    exit_status=$?
+    if [[ $exit_status != 0 || "${chosen_uat_decoder}" == "None" ]]; then
+        install_adsb_decoder="false"
     fi
 fi
 
@@ -50,12 +71,22 @@ function install_dump1090_fa() {
     fi
 }
 
+function install_readsb() {
+    chmod +x ${RECEIVER_BASH_DIRECTORY}/decoders/readsb.sh
+    ${RECEIVER_BASH_DIRECTORY}/decoders/readsb.sh
+    if [[ $? != 0 ]] ; then
+        exit 1
+    fi
+}
+
 
 ## UAT DECODERS
 
-# Flightaware dump978
+uat_decoder_installed="false"
 install_uat_decoder="false"
+
 if [[ $(dpkg-query -W -f='${STATUS}' dump978-fa 2>/dev/null | grep -c "ok installed") == 1 ]]; then
+    uat_decoder_installed="true"
     chosen_uat_decoder="dump978-fa"
     if [[ $(sudo dpkg -s dump978-fa 2>/dev/null | grep -c "Version: ${dump978_fa_current_version}") == 0 ]]; then
         whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
@@ -74,15 +105,20 @@ if [[ $(dpkg-query -W -f='${STATUS}' dump978-fa 2>/dev/null | grep -c "ok instal
             install_uat_decoder="true"
         fi
     fi
-else
-    whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
-             --title "FlightAware dump978" \
-             --defaultno \
-             --yesno "FlightAware dump978 is capable of demodulating UAT received by an SDR device.\n\nGitHub Repository: https://github.com/flightaware/dump978\n\nWould you like to install FlightAware dump978?" \
-             10 65
-    if [[ $? == 0 ]]; then
-        install_uat_decoder="true"
-        chosen_uat_decoder="dump978-fa"
+fi
+
+if [[ "${uat_decoder_installed}" == "false" ]]; then
+    install_uat_decoder="true"
+    chosen_uat_decoder=$(whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
+                                  --title "UAT Decoder Selection" \
+                                  --menu "The following UAT decoders are available for installation." \
+                                  16 100 9 \
+                                  "None" "Do not install a UAT decoder." \
+                                  "dump978-fa" "FlightAware's version of the dump978 decoder." \
+                                  3>&2 2>&1 1>&3)
+    exit_status=$?
+    if [[ $exit_status != 0 || "${chosen_uat_decoder}" == "None" ]]; then
+        install_uat_decoder="false"
     fi
 fi
 
@@ -94,11 +130,14 @@ function install_dump978_fa() {
     fi
 }
 
+
 ## ACARS DECODERS
 
-# ACARSDEC
+acars_decoder_installed="false"
 install_acars_decoder="false"
+
 if [[ -f /etc/systemd/system/acarsdec.service ]]; then
+    acars_decoder_installed="true"
     chosen_acars_decoder="acarsdec"
     whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
              --title "Reinstall ACARSDEC Decoder" \
@@ -108,21 +147,90 @@ if [[ -f /etc/systemd/system/acarsdec.service ]]; then
     if [[ $? == 0 ]]; then
         install_acars_decoder="true"
     fi
-else
-    whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
-             --title "ACARSDEC Decoder" \
-             --defaultno \
-             --yesno "ACARSDEC is a multi-channels acars decoder with built-in rtl_sdr, airspy front end or sdrplay device.\n\nGitHub Repository: https://github.com/TLeconte/acarsdec\n\nWould you like to install ACARSDEC?" \
-             10 65
-    if [[ $? == 0 ]]; then
-        install_acars_decoder="true"
-        chosen_acars_decoder="acarsdec"
+fi
+
+if [[ "${acars_decoder_installed}" == "false" ]]; then
+    install_acars_decoder="true"
+    chosen_acars_decoder=$(whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
+                                    --title "ACARS Decoder Selection" \
+                                    --menu "The following ACARS decoders are available for installation." \
+                                    16 100 9 \
+                                    "None" "Do not install an ACARS decoder." \
+                                    "acarsdec" "Acarsdec is a multi-channels acars decoder." \
+                                    3>&2 2>&1 1>&3)
+    exit_status=$?
+    if [[ $exit_status != 0 || "${chosen_acars_decoder}" == "None" ]]; then
+        install_acars_decoder="false"
     fi
 fi
 
 function install_acarsdec() {
     chmod +x ${RECEIVER_BASH_DIRECTORY}/decoders/acarsdec.sh
     ${RECEIVER_BASH_DIRECTORY}/decoders/acarsdec.sh
+    if [[ $? != 0 ]] ; then
+        exit 1
+    fi
+}
+
+
+## VDL MODE 2 DECODERS
+
+vdlm2_decoder_installed="false"
+install_vdlm2_decoder="false"
+
+if [[ -f /etc/systemd/system/dumpvdl2.service ]]; then
+    vdlm2_decoder_installed="true"
+    chosen_vdlm2_decoder="dumpvdl2"
+    whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
+             --title "Reinstall dumpvdl2 Decoder" \
+             --defaultno \
+             --yesno "The option to rebuild and reinstall dumpvdl2 is available.\n\nWould you like to rebuild and reinstall dumpvdl2?" \
+             9 65
+    if [[ $? == 0 ]]; then
+        install_vdlm2_decoder="true"
+    fi
+fi
+
+if [[ -f /etc/systemd/system/vdlm2dec.service ]]; then
+    vdlm2_decoder_installed="true"
+    chosen_vdlm2_decoder="vdlm2dec"
+    whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
+             --title "Reinstall VDLM2DEC Decoder" \
+             --defaultno \
+             --yesno "The option to rebuild and reinstall VDLM2DEC is available.\n\nWould you like to rebuild and reinstall VDLM2DEC?" \
+             9 65
+    if [[ $? == 0 ]]; then
+        install_vdlm2_decoder="true"
+    fi
+fi
+
+if [[ "${vdlm2_decoder_installed}" == "false" ]]; then
+    install_vdlm2_decoder="true"
+    chosen_vdlm2_decoder=$(whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
+             --title "VLD Mode 2 Decoder Selection" \
+             --menu "The following VLD Mode 2 decoders are available for installation." \
+             16 100 9 \
+             "None" "Do not install a VLD decoder." \
+             "vdlm2dec" "vdlm2dec is a VDL Mode 2 decoder." \
+             "dumpvdl2" "dumpvdl2 is a VDL Mode 2 message decoder." \
+             3>&2 2>&1 1>&3)
+    exit_status=$?
+    if [[ $exit_status != 0 || "${chosen_vdlm2_decoder}" == "None" ]]; then
+        install_vdlm2_decoder="false"
+    fi
+fi
+
+function install_dumpvdl2() {
+    chmod +x ${RECEIVER_BASH_DIRECTORY}/decoders/dumpvdl2.sh
+    ${RECEIVER_BASH_DIRECTORY}/decoders/dumpvdl2.sh
+    if [[ $? != 0 ]] ; then
+        exit 1
+    fi
+}
+
+function install_vdlm2dec() {
+    chmod +x ${RECEIVER_BASH_DIRECTORY}/decoders/vdlm2dec.sh
+    ${RECEIVER_BASH_DIRECTORY}/decoders/vdlm2dec.sh
     if [[ $? != 0 ]] ; then
         exit 1
     fi
@@ -338,10 +446,8 @@ function install_beastsplitter() {
 
 # Duck DNS
 if [[ ! -f "${RECEIVER_BUILD_DIRECTORY}/duckdns/duck.sh" ]]; then
-    # Duck DNS does not appear to be set up on this device.
     extras_list=("${extras_list[@]}" 'Duck DNS Free Dynamic DNS Hosting' '' OFF)
 else
-    # Offer the option to install/setup Duck DNS once more.
     extras_list=("${extras_list[@]}" 'Duck DNS Free Dynamic DNS Hosting (reinstall)' '' OFF)
 fi
 
@@ -349,6 +455,36 @@ function install_duckdns() {
     chmod +x ${RECEIVER_BASH_DIRECTORY}/extras/duckdns.sh
     ${RECEIVER_BASH_DIRECTORY}/extras/duckdns.sh
     if [[ $? -ne 0 ]] ; then
+        exit 1
+    fi
+}
+
+# Graphs1090
+if [[ ! -f /lib/systemd/system/graphs1090.service ]]; then
+    extras_list=("${extras_list[@]}" 'Graphs1090' '' OFF)
+else
+    extras_list=("${extras_list[@]}" 'Graphs1090 (reinstall)' '' OFF)
+fi
+
+function install_graphs1090() {
+    chmod +x ${RECEIVER_BASH_DIRECTORY}/extras/graphs1090.sh
+    ${RECEIVER_BASH_DIRECTORY}/extras/graphs1090.sh
+    if [[ $? != 0 ]] ; then
+        exit 1
+    fi
+}
+
+# tar1090
+if [[ ! -f /lib/systemd/system/tar1090.service ]]; then
+    extras_list=("${extras_list[@]}" 'tar1090' '' OFF)
+else
+    extras_list=("${extras_list[@]}" 'tar1090 (reinstall)' '' OFF)
+fi
+
+function install_tar1090() {
+    chmod +x ${RECEIVER_BASH_DIRECTORY}/extras/tar1090.sh
+    ${RECEIVER_BASH_DIRECTORY}/extras/tar1090.sh
+    if [[ $? != 0 ]] ; then
         exit 1
     fi
 }
@@ -365,7 +501,7 @@ whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
 
 declare confirmation_message
 
-if [[ "${install_adsb_decoder}" == "false" && "${install_uat_decoder}" == "false" && "${install_acars_decoder}" == "false" && "${install_portal}" == "false" && ! -s "${RECEIVER_ROOT_DIRECTORY}/FEEDER_CHOICES" && ! -s "${RECEIVER_ROOT_DIRECTORY}/EXTRAS_CHOICES" ]]; then
+if [[ "${install_adsb_decoder}" == "false" && "${install_uat_decoder}" == "false" && "${install_acars_decoder}" == "false" && "${install_vdlm2_decoder}" == "false" && "${install_portal}" == "false" && ! -s "${RECEIVER_ROOT_DIRECTORY}/FEEDER_CHOICES" && ! -s "${RECEIVER_ROOT_DIRECTORY}/EXTRAS_CHOICES" ]]; then
     whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
              --title "Nothing to be done" \
              --msgbox "Nothing has been selected to be installed so the script will exit now." \
@@ -382,6 +518,9 @@ else
         case ${chosen_adsb_decoder} in
             "dump1090-fa")
                 confirmation_message="${confirmation_message}\n  * FlightAware dump1090"
+                ;;
+            "readsb")
+                confirmation_message="${confirmation_message}\n  * Readsb"
                 ;;
         esac
     fi
@@ -400,6 +539,18 @@ else
         case ${chosen_acars_decoder} in
             "acarsdec")
                 confirmation_message="${confirmation_message}\n  * ACARSDEC"
+                ;;
+        esac
+    fi
+
+    # VDL Mode 2 decoders
+    if [[ "${install_vdlm2_decoder}" = "true" ]]; then
+        case ${chosen_vdlm2_decoder} in
+            "dumpvdl2")
+                confirmation_message="${confirmation_message}\n  * dumpvdl2"
+                ;;
+            "vdlm2dec")
+                confirmation_message="${confirmation_message}\n  * vdlm2dec"
                 ;;
         esac
     fi
@@ -445,6 +596,9 @@ if [[ "${install_adsb_decoder}" == "true" ]]; then
         "dump1090-fa")
              install_dump1090_fa
              ;;
+        "readsb")
+             install_readsb
+             ;;
     esac
 fi
 
@@ -463,6 +617,18 @@ if [[ "${install_acars_decoder}" == "true" ]]; then
         "acarsdec")
              install_acarsdec
              ;;
+    esac
+fi
+
+# VDL Decoders
+if [[ "${install_vdlm2_decoder}" == "true" ]]; then
+    case ${chosen_vdlm2_decoder} in
+        "dumpvdl2")
+            install_dumpvdl2
+            ;;
+        "vdlm2dec")
+            install_vdlm2dec
+            ;;
     esac
 fi
 
@@ -552,6 +718,12 @@ if [[ -s "${RECEIVER_ROOT_DIRECTORY}/EXTRAS_CHOICES" ]]; then
             "Duck DNS Free Dynamic DNS Hosting"|"Duck DNS Free Dynamic DNS Hosting (reinstall)")
                 run_duckdns_script="true"
                 ;;
+            "Graphs1090"|"Graphs1090 (reinstall)")
+                run_graphs1090_script="true"
+                ;;
+            "tar1090"|"tar1090 (reinstall)")
+                run_tar1090_script="true"
+                ;;
         esac
     done < ${RECEIVER_ROOT_DIRECTORY}/EXTRAS_CHOICES
 fi
@@ -562,6 +734,14 @@ fi
 
 if [[ "${run_duckdns_script}" == "true" ]]; then
     install_duckdns
+fi
+
+if [[ "${run_graphs1090_script}" == "true" ]]; then
+    install_graphs1090
+fi
+
+if [[ "${run_tar1090_script}" == "true" ]]; then
+    install_tar1090
 fi
 
 exit 0