+++ install.sh
@@ -1,47 +1,13 @@
 #!/bin/bash
 
-#####################################################################################
-#                                   ADS-B RECEIVER                                  #
-#####################################################################################
-#                                                                                   #
-#  A set of scripts created to automate the process of installing the software      #
-#  needed to setup an ADS-B/Mode S decoder as well as feeders which are capable of  #
-#  sharing your ADS-B results with many of the most popular ADS-B aggregate sites.  #
-#                                                                                   #
-#  Project Hosted On GitHub: https://github.com/jprochazka/adsb-receiver            #
-#                                                                                   #
-#  -------------------------------------------------------------------------------  #
-#                                                                                   #
-# Copyright (c) 2015 Joseph A. Prochazka and Contributors                           #
-#                                                                                   #
-# Permission is hereby granted, free of charge, to any person obtaining a copy      #
-# of this software and associated documentation files (the "Software"), to deal     #
-# in the Software without restriction, including without limitation the rights      #
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell         #
-# copies of the Software, and to permit persons to whom the Software is             #
-# furnished to do so, subject to the following conditions:                          #
-#                                                                                   #
-# The above copyright notice and this permission notice shall be included in all    #
-# copies or substantial portions of the Software.                                   #
-#                                                                                   #
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR        #
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,          #
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE       #
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER            #
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,     #
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE     #
-# SOFTWARE.                                                                         #
-#                                                                                   #
-#####################################################################################
-
-
 ## ASSIGN VARIABLE
 
-project_version="2.8.6"
+project_version="2.8.7"
 
 printf -v date_time '%(%Y-%m-%d_%H-%M-%S)T' -1
 log_file="adsb-installer_${date_time}.log"
 logging_enabled="true"
+project_branch="master"
 
 
 ## FUNCTIONS
@@ -128,8 +94,6 @@ export RECEIVER_MTA=$mta
 
 ## SET PROJECT VARIABLES
 
-project_branch="master"
-
 export RECEIVER_PROJECT_TITLE="ADS-B Receiver Installer v${project_version}"
 export RECEIVER_ROOT_DIRECTORY=$PWD
 export RECEIVER_BASH_DIRECTORY=$PWD/bash
@@ -193,6 +157,17 @@ export RECEIVER_OS_RELEASE=`. /etc/os-release; echo ${VERSION_ID/*, /}`
 export RECIEVER_CPU_ARCHITECTURE=`uname -m | tr -d "\n\r"`
 export RECEIVER_CPU_REVISION=`grep "^Revision" /proc/cpuinfo | awk '{print $3}'`
 
+## INIT DECODER DEVICE ASSIGNMENT VARIABLES
+
+export RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER
+export RECEIVER_DEVICE_ASSIGNED_TO_ADSB_DECODER
+export RECEIVER_DEVICE_ASSIGNED_TO_UAT_DECODER
+export RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER
+export RECEIVER_ACARS_DECODER_SOFTWARE
+export RECEIVER_ADSB_DECODER_SOFTWARE
+export RECEIVER_UAT_DECODER_SOFTWARE
+export RECEIVER_VDLM2_DECODER_SOFTWARE
+
 
 ## EXECUTE BASH/INIT.SH
 
@@ -217,6 +192,14 @@ unset RECEIVER_OS_DISTRIBUTION
 unset RECEIVER_OS_RELEASE
 unset RECIEVER_CPU_ARCHITECTURE
 unset RECEIVER_CPU_REVISION
+unset RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER
+unset RECEIVER_DEVICE_ASSIGNED_TO_ADSB_DECODER
+unset RECEIVER_DEVICE_ASSIGNED_TO_UAT_DECODER
+unset RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER
+unset RECEIVER_ACARS_DECODER_SOFTWARE
+unset RECEIVER_ADSB_DECODER_SOFTWARE
+unset RECEIVER_UAT_DECODER_SOFTWARE
+unset RECEIVER_VDLM2_DECODER_SOFTWARE
 unset RECEIVER_PROJECT_BRANCH
 unset RECEIVER_DEVELOPMENT_MODE
 unset RECEIVER_LOGGING_ENABLED